fixed all unit tests for java/jarfile
all unit tests explicitly set [minecraft] profile and [java] jarfile in
the test itself, rather than letting it get set by mc.start().  The
assumption that the profile would be 1.7.9 was never meant to be
permanent and a profile itself is not required for operation. [java]
jarfile, on the other hand, is required.