server still can be started without valid profile
if invalid profile (dir missing in filesystem) in server.config ,c
opy_profile subsequently skipped and mc.start continues.

this will be the case for many converted webuis where the vanilla179
name will not match the 1.7.9 file naming over the new webui.

It will pop up the 'choose a profile/jar' modal instead.  Profile can
still be ignored, but jar may not be.