re-revised profile downloading
the server backend will re-download ALL profile types upon each
successful download; this is the default behavior.

even though this means (currently) 5 HTTP get requests and a lot of
parsing, it's also the only way the 'downloaded' confirmation shows on
the profile list.

For the time being, this excess should be manageable.