Prepare for broadcasting from multiple IPs
This slightly rewrites the broadcast loop so that it checks if there is a socket for an IP and if not, creates it. By hardcoding '0.0.0.0' the functionality is equal to before this change and all that is needed is to determine each Minecraft server's IP, but I haven't yet figured out how.