blacklisted various directories from fireworm
the biggest concern of performance and operability was getting fireworm
around blocking I/O from high file counts.  While this does not fix the
issue in a strict sense, it alleviates the issue when it is caused by
dynmap and excessively large worlds by making ensuring fireworm does not
trigger on files that need not be watched in the first place.