added client-side rate limiting to logs
note, the transmission of logs from server-> client/s will continue.
this makes the potential impact on the server-side relatively benign.
rate limiting occurs at 50+ lines per second.

at any time the logs produce more than this, it is unlikely it will be
as useful to have live tailing in the webui than it is to open the logs
directly in a terminal.